test: better I/O tests (evtmax=-1, nprocs=-1)
I/O with -nprocs=-1 and -evtmax=-1 used to dead lock.
This was because the task producing the first data items was "lost"
among the other tasks.
When that input stream was hiting the io.EOF of the underlying
io.Reader, other workers may have already been waiting on the
"synthetic" chan int64 evts used to prime all workers and tasks.

The new setup is priming the whole graph of tasks by having a single
app.input-stream task, reading all input data and sending it over a
'chan context' events.
Each gang of workers receive from that channel.
That channel is closed when io.EOF is encountered.

In the case of an application with no input-file (e.g. a Monte-Carlo
generator job) a dummy input-stream is created.