hbook/ntuple: make ScanH1D less memory hungry
This CL makes Ntuple.ScanH1D less memory hungry.
If the 1D-histogram wasn't provided, Scan1D used to go through the
n-tuple data, calculate xmin and xmax, and populate a slice with the
wanted data.
This resulted in a (possibly) big data slice which would then be
discarded (once it was used to fill the histogram.)

Instead, crawl through the whole n-tuple data to calculate xmin and
xmax.
Create the histogram with the computed interval and then re-crawl
through the whole n-tuple data, filling the histogram along the way.