hbook: migrate to YODA-like H1D implementation
This CL migrates to a YODA-like implementation for H1D (similar to what
had already been done for H2D) and adds YODA I/O for H1D as well.

This CL also leverages the new go-hep/brio package to automatically
generate serializers (implementing encoding.Binary(Unm|M)arshaler).

The API of H1D slightly changed to be closer to YODA::Histo1D and to not
be confused with H2D:
 H1D.Mean() -> H1D.XMean()
 H2D.MeanX() -> H2D.XMean()
 H2D.MeanY() -> H2D.YMean()

Fixes #13.
Fixes #18.