rootio: correctly update firstEntry field in tbranch
When a Tree is used multiple times, its branches carry out state from
traversal to traversal.
It happened that the firstEntry field of a tbranch could be carried out
from a previous traversal and e.g. had the entry value of say the last
basket.
It would result in a negative offset and then an error.

Fixes go-hep/hep#48.