hplot: correctly compute DataRange for S2D with error bars
Previously, the DataRange implementation for S2D plots didn't take into
account the data extend coming from the X and/or Y error bars.

This CL fixes this by first calling the underlying DataRanger for the
points and then increases the extend for the X and Y error bars, if any.