hplot: Changed the way HInfoStyles work, and added HInfoStdDev
Previously, there seemed to be an incorrectly configured set of
constants enumerating the different styles, and incomplete code for
adding info to the legend.  The enumeration has been changed to be
bitset-like, and the code that adds the info to the legend now iterates
over the bits, allowing the bitwise ORing of styles.

Additionally, HInfoStdDev has been added to the style enumeration, and
it replaces HInfoRMS as one of the styles OR'd for HInfoSummary.

Finally, HInfoStyle has been changed from int to uint32.