xrootd: add protocol request
Protocol request is used for obtaining the protocol version number,
type of server and possible security requirements.

These security requirements determine which requests should be
signed. This info is provided to Client via protocol.SignRequirements.
The Client will use sr.Needed to determine if the request should be
signed via Sigver request.

Actual sign requirements depending on the security level will be
added as part of the request implementation in the following form:
 if level >= protocol.Pedantic {
     sr.requirements[dirlist.RequestID] = protocol.SignNeeded
     // ...
 }

Updates go-hep/hep#170.