xrootd/client: fix test
If cancel() is called before client handled request, then the client
will raise an error indicating that (expected behaviour if we called
Close while some requests were blocked).

In case of test failure, it's safe to call cancel()
before the end of request handling because of the fact that
the actual error message will still be present at top of the log.