rootio: reduce number of count-branch load in scanners
Previously, whenever an activated branch had a dependency (e.g. a
LeafCount), we added that dependency.
The dependency branch would then be loaded and read for every call to
Scanner.Scan.

If branch-1 and branch-2 depended on branch-3, branch-3 would be loaded
twice.

This CL modifies the Scanners to only add depended branches once.