xrootd/xrdproto/auth/krb5: handle unsupported directive parsing errors
This CL leverages gokrb5.v6 to properly handle unsupported directive
errors that stem from e.g. krb5.conf sections like:

```
[realms]
 CERN.CH  = {
  default_domain = cern.ch
  kpasswd_server = cerndc.cern.ch
  admin_server = cerndc.cern.ch
   kdc = cerndc.cern.ch
    v4_name_convert = {
     host = {
        rcmd = host
     }
   }
 }
```

the v4 parts are not supported by gokrb5.
this used to be fatal (in the sense that xrootd couldn't generate a
proper authentication token) but didn't need to be.

now, gokrb5.v6 returns this new config.UnsupportedDirective error that
we recognize and "handle" (by ignoring it.)

Fixes go-hep/hep#355.