groot/rdict: remove init-based code
This CL removes the init-based code used to locate a few packages
(encoding, rbytes, ...) that are needed during code generation.
This required to have a Go installation and the source code of groot
whenever a binary importing groot or groot/rdict was launched.

Delay this code to be executed during the creation of an
rdict.Generator.