groot/cmd/root-gen-rfunc: first import
This CL adds a new command to automatically generate the rfunc.Formula
implementation corresponding to a func signature (or a function in a
package (ex: math.Abs)).

This CL also drops the automagical registration of rfunc.Formula
implementations from the rfunc package (as it would prevent the linker
from applying dead code elimination).

Fixes #729.