groot/riofs: fix overlaping blocks w/ big-file mode
TKeys sizes computation was broken for "big files".
The code was indeed changing the `key.rvers` field to indicate
the "big file" mode was enabled, but it wasn't updating the
on-disk size of the TKey (ie: missing 8bytes).

This CL computes the correct size in one-go instead of fixing it
after the facts.

Fixes go-hep/hep#821.