groot/rdict: proper handling of 'This' streamers
'This' streamers are used to extract the type argument of templates.
e.g.:
 class MyVector: public std::vector<Foo> {...};
would be encoded as a TStreamerInfo named 'MyVector' and an element
named "This" with a title "<Foo> Used to call ...".

This technique seems to be used to properly extract 'Foo'.
This seems to be also applied for other containers. (vector, deque, ...)