brio-gen: cleanup files produced by test on test success
In tests of brio-gen, when a failure occurs (generated output changes)
the new version is kept on disk. This is the existing behaviour.

With this patch we remove the new version if there is no difference
(test success). This allows to automatically remove those temporary
files that are useful only for debugging a test failure.