groot/cmd/root-gen-streamer: use an in-memory buffer for generation
This CL modifies root-gen-streamer to use an in-memory buffer as the
destination of the code generation.
This is needed because the Go-1.19.10 version of x/go/packages.Load will
choke on an empty "output_gen.go" file.
This was the case as the old version root-gen-streamer was creating the
destination file for the code generation before calling the streamer
generation.

Signed-off-by: Sebastien Binet <binet@cern.ch>